nbd-server: handle auth for v6-mapped IPv4 addresses
This patch adds support for using IPv4 addresses in the nbd-server auth file
even if the socket provides v6-mapped addresses. It extends the comparison
code in address_matches to handle the two cases where the auth file and the
socket use different address families.

This fixes issue #35 in the github repository.